Refactor accordion
